@@ -11,7 +11,7 @@
         }
     ],
     "require": {
-        "php": "^7.3 || ^8.0"
+        "php": "^7.4 || ^8.0"
     },
     "require-dev": {
         "bamarni/composer-bin-plugin": "^1.4",
@@ -45,7 +45,7 @@
         "cs-fix": "./vendor/bin/phpcbf src tests",
         "metrics": "./vendor/bin/phpmetrics --report-html=build/metrics --exclude=Exception src",
         "clean": ["./vendor/bin/phpstan clear-result-cache", "./vendor/bin/psalm --clear-cache"],
-        "sa": ["./vendor/bin/phpstan analyse -c phpstan.neon", "psalm --show-info=true"],
+        "sa": ["./vendor/bin/phpstan analyse -c phpstan.neon", "./vendor/bin/psalm --show-info=true"],
         "tests": ["@cs", "@sa", "@test"],
         "build": ["@clean", "@cs", "@sa", "@pcov", "@metrics"]
     },
@@ -61,5 +61,10 @@
         "metrics": "Build metrics report",
         "tests": "Run tests and quality checks",
         "build": "Build project"
+    },
+    "config": {
+        "allow-plugins": {
+            "bamarni/composer-bin-plugin": true
+        }
     }
 }
